requirements.txt
torch>=2.0.0
torchvision>=0.15.1
matplotlib>=3.7.1
numpy>=1.24.3
scikit-learn>=1.2.2
tqdm>=4.65.0
pillow>=9.5.0
seaborn>=0.12.2
pandas>=1.5.3
timm>=0.9.2
peft>=0.5.0
transformers>=4.31.0
pathlib>=1.0.1 